+++ src/app.rs
@@ -6,64 +6,7 @@ use yew_router::prelude::*;
 use crate::components::body::*;
 use crate::components::header::*;
 use crate::components::meta::*;
-use crate::components::search_field::*;
-use crate::pages::*;
 use crate::utils::*;
-
-#[derive(Routable, PartialEq, Eq, Clone, Debug)]
-pub enum Route {
-    #[at("/post/new")]
-    NewPost,
-    #[at("/post/edit/:id")]
-    EditPost { id: u64 },
-    #[at("/post/:slug/:id")]
-    Post { slug: String, id: u64 },
-    #[at("/posts")]
-    Posts,
-    #[at("/posts/search")]
-    PostsSearchRoot,
-    #[at("/posts/search/:query")]
-    PostsSearch { query: String },
-    #[at("/author/:slug")]
-    Author { slug: String },
-    #[at("/authors")]
-    Authors,
-    #[at("/authors/search")]
-    AuthorsSearchRoot,
-    #[at("/authors/search/:query")]
-    AuthorsSearch { query: String },
-    #[at("/")]
-    Home,
-    #[not_found]
-    #[at("/404")]
-    NotFound,
-}
-
-impl Route {
-    pub fn switch(route: Route) -> Html {
-        match route {
-            Route::NewPost => html! { <EditPost id={ None } /> },
-            Route::EditPost { id } => html! { <EditPost { id } /> },
-            Route::Post { slug, id } => html! { <Post { slug } { id } /> },
-            Route::Home | Route::Posts => html! { <Posts />},
-            Route::PostsSearchRoot => {
-                html! { <Search mode={ SearchMode::Posts { query: None } } />}
-            }
-            Route::PostsSearch { query } => {
-                html! { <Search mode={ SearchMode::Posts { query: Some(query) } } />}
-            }
-            Route::Author { slug } => html! { <Author { slug } /> },
-            Route::Authors => html! { <Authors /> },
-            Route::AuthorsSearchRoot => {
-                html! { <Search mode={ SearchMode::Authors { query: None } } />}
-            }
-            Route::AuthorsSearch { query } => {
-                html! { <Search mode={ SearchMode::Authors { query: Some(query) } } />}
-            }
-            Route::NotFound => html! { <PageNotFound /> },
-        }
-    }
-}
 
 #[derive(Properties, PartialEq, Debug)]
 struct MainProps {
